refactor: 영화 상세 정보 메서드 MovieCard -> MovieCardList
